internal: tag tests follow their vocabularies into the command test files
The vocabularies moved to checks_changelog.go and checks_coverage.go;
their tests now mirror them, per the test-files-mirror-source
convention. tags_test.go is removed: the dimension mechanism is
exercised through both vocabularies' sweeps and the OR-semantics test.
